[#3688] Adapt tests
